VERY IMPORTANT UPDATE: Moved CFG variables to DB. Removed PHP notices.

+++ cron/receive_bitcoin.php
@@ -2,7 +2,7 @@
 <?php
 echo "Beginning Receive Bitcoin processing...".PHP_EOL;
 
-include 'cfg.php';
+include 'common.php';
 
 $CFG->session_active = true;
 $transactions_dir = $CFG->dirroot.'transactions/';
@@ -75,11 +75,11 @@ foreach ($transactions as $t_id) {
 		continue;
 	
 	$transaction = $bitcoin->gettransaction($t_id);
-	if ($transaction_log[$t_id]) {
+	if (!empty($transaction_log[$t_id])) {
 		unlink($transactions_dir.$t_id);
 		continue;
 	}
-	if (!$transaction['details'])
+	if (empty($transaction['details']))
 		continue;
 	
 	//$raw = $bitcoin->decoderawtransaction($bitcoin->getrawtransaction($t_id));
@@ -91,16 +91,16 @@ foreach ($transactions as $t_id) {
 	
 	foreach ($transaction['details'] as $detail) {
 		if ($detail['category'] == 'receive') {
-			$user_id = $addresses[$detail['address']];
-			$request_id = $requests[$transaction['txid']];
+			$user_id = (!empty($addresses[$detail['address']])) ? $addresses[$detail['address']] : false;
+			$request_id = (!empty($requests[$transaction['txid']])) ? $requests[$transaction['txid']] : false;
 			
-			if ($system[$detail['address']] == 'Y') {
+			if (!empty($system[$detail['address']]) && $system[$detail['address']] == 'Y') {
 				if ($transaction['confirmations'] > 0) {
 					$hot_wallet_in = $detail['amount'];
 				}
 				continue;
 			}
-			elseif ($system[$detail['address']] == 'N') {
+			elseif (!empty($system[$detail['address']]) && $system[$detail['address']] == 'N') {
 				unlink($transactions_dir.$t_id);
 				break;
 			}
